refactor(e2e): extract shared receipt-fixture helpers from admin + payments specs
Both apps' new "receipt cell exposes a working signed URL" tests had
identical setup (1px PNG constant, uploadTestReceipt over the Supabase
storage REST endpoint, PATCH order.receipt_url) and an identical
assertion body (≈30 lines verifying href shape + image fetch).

Moves all of it into apps/auth/e2e/helpers/receiptFixtures.ts (the
established home for cross-app e2e helpers — sibling to session.ts,
cleanup.ts, constants.ts):

  - ONE_PX_PNG_BASE64 / RECEIPT_FILENAME constants
  - uploadTestReceipt(storagePath, png?)
  - patchOrderReceiptUrl(orderId, storagePath)
  - verifyReceiptLinkResolves(page, { testIdPrefix, orderId })

Net: 21 fewer lines, single point of maintenance, and any future app
that adds a similar regression test (studio sales reports, etc.) can
reuse the fixture instead of copying.
