+++ apps/payments/e2e/seller-reports.spec.ts
@@ -15,6 +15,12 @@ import {
   supabaseAdmin,
   type TestUser,
 } from "../../auth/e2e/helpers/session";
+import {
+  RECEIPT_FILENAME,
+  patchOrderReceiptUrl,
+  uploadTestReceipt,
+  verifyReceiptLinkResolves,
+} from "../../auth/e2e/helpers/receiptFixtures";
 
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 const { resolveE2EAppUrls } = require(
@@ -41,46 +47,6 @@ const TEST_ORDER_ITEM = {
   unit_price: 25000,
   currency: "COP",
 };
-
-// 1x1 transparent PNG so the uploaded receipt is a real image the
-// browser can render when the signed URL resolves.
-const ONE_PX_PNG_BASE64 =
-  "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mNkYAAAAAYAAjCB0C8AAAAASUVORK5CYII=";
-const RECEIPT_FILENAME = "receipt-e2e.png";
-
-async function uploadTestReceipt(
-  storagePath: string,
-  pngBase64: string,
-): Promise<void> {
-  const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-  const serviceRoleKey = process.env.SUPABASE_SERVICE_ROLE_KEY;
-  if (!supabaseUrl || !serviceRoleKey) {
-    throw new Error(
-      "NEXT_PUBLIC_SUPABASE_URL and SUPABASE_SERVICE_ROLE_KEY are required for e2e receipt upload.",
-    );
-  }
-
-  const response = await fetch(
-    `${supabaseUrl}/storage/v1/object/receipts/${storagePath}`,
-    {
-      method: "POST",
-      headers: {
-        apikey: serviceRoleKey,
-        Authorization: `Bearer ${serviceRoleKey}`,
-        "Content-Type": "image/png",
-        "x-upsert": "true",
-      },
-      body: Buffer.from(pngBase64, "base64"),
-    },
-  );
-
-  if (!response.ok) {
-    const body = await response.text();
-    throw new Error(
-      `Failed to upload e2e receipt file: ${response.status} ${body}`,
-    );
-  }
-}
 
 // ─── Test suite ───────────────────────────────────────────────────
 
@@ -119,31 +85,8 @@ test.describe.serial("Seller Reports page", () => {
     // Upload a receipt image and attach its storage path to the order so
     // the report can verify the API converts the path to a signed URL.
     receiptStoragePath = `${orderId}/${RECEIPT_FILENAME}`;
-    await uploadTestReceipt(receiptStoragePath, ONE_PX_PNG_BASE64);
-
-    const supabaseUrl = process.env.NEXT_PUBLIC_SUPABASE_URL;
-    const serviceRoleKey = process.env.SUPABASE_SERVICE_ROLE_KEY;
-    if (!supabaseUrl || !serviceRoleKey) {
-      throw new Error(
-        "NEXT_PUBLIC_SUPABASE_URL and SUPABASE_SERVICE_ROLE_KEY are required for e2e setup.",
-      );
-    }
-    const patchResponse = await fetch(
-      `${supabaseUrl}/rest/v1/orders?id=eq.${orderId}`,
-      {
-        method: "PATCH",
-        headers: {
-          apikey: serviceRoleKey,
-          Authorization: `Bearer ${serviceRoleKey}`,
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ receipt_url: receiptStoragePath }),
-      },
-    );
-    if (!patchResponse.ok) {
-      const body = await patchResponse.text();
-      throw new Error(`Failed to set order receipt_url: ${body}`);
-    }
+    await uploadTestReceipt(receiptStoragePath);
+    await patchOrderReceiptUrl(orderId, receiptStoragePath);
 
     await adminInsert("order_items", {
       order_id: orderId,
@@ -326,27 +269,10 @@ test.describe.serial("Seller Reports page", () => {
       timeout: ELEMENT_TIMEOUT_MS,
     });
 
-    const receiptCell = page.getByTestId(
-      `seller-report-row-receipt-${orderId}`,
-    );
-    await expect(receiptCell).toBeVisible({ timeout: MUTATION_WAIT_MS });
-
-    const receiptLink = receiptCell.getByRole("link");
-    await expect(receiptLink).toBeVisible({ timeout: ELEMENT_TIMEOUT_MS });
-
-    const href = await receiptLink.getAttribute("href");
-    expect(href).toBeTruthy();
-    // The href must be a fully-qualified URL — not the raw storage path —
-    // and must be a Supabase signed URL for the receipt object.
-    expect(href).toMatch(/^https?:\/\//);
-    expect(href).toContain(`/storage/v1/object/sign/receipts/${orderId}/`);
-    expect(href).toContain("token=");
-
-    // The signed URL must actually resolve to an image (the bug was that the
-    // raw storage path was rendered, producing a relative URL that 404'd).
-    const fetchedReceipt = await page.request.get(href as string);
-    expect(fetchedReceipt.ok()).toBe(true);
-    expect(fetchedReceipt.headers()["content-type"]).toContain("image");
+    await verifyReceiptLinkResolves(page, {
+      testIdPrefix: "seller-report-row-receipt",
+      orderId,
+    });
   });
 
   test("status=pending filter hides the approved seeded order", async ({
